+++ bin/autobot.dart
@@ -1,3 +1,4 @@
 import 'package:autobot/autobot.dart' as autobot;
 
+// Runs autobot with the given cli arguments.
 void main(List<String> arguments) => autobot.main(arguments);