+++ lib/commands/run/input_reader.dart
@@ -3,12 +3,14 @@ part of 'run.dart';
 typedef Key = String;
 typedef Value = String;
 
+/// Reads the input argument.
 class InputReader {
   InputReader(this.owner);
 
   final RunCommand owner;
   final _argumentInputs = <Key, Value>{};
 
+  /// Reads the input argument and collects all variables.
   InputReader collectInputsFromArgs() {
     final argInputs = owner.argResults![owner.kOptionInput] ?? const [];
 
@@ -24,14 +26,14 @@ class InputReader {
     return this;
   }
 
-  List<Input> askForInputvalues(TemplateDef template) {
-    return template.inputs.map((inputDef) {
-      final argumentInput = _argumentInputs.tryGet(inputDef.key);
-
-      return Input(
-        key: inputDef.key,
-        value: argumentInput ?? ask(yellow(inputDef.prompt)),
-      );
-    }).toList();
+  /// Asks the cli user for the missing inputs value and returns them as a [Map].
+  Map<Key, Value> askForInputvalues(TemplateDef template) {
+    return template.inputs.toMap(
+      keyProvider: (inputDef) => inputDef.key,
+      valueProvider: (inputDef) {
+        final argumentInput = _argumentInputs.tryGet(inputDef.key);
+        return argumentInput ?? ask(yellow(inputDef.prompt));
+      },
+    );
   }
 }