+++ lib/commands/run/template_reader.dart
@@ -6,6 +6,7 @@ class RunTemplateReader {
   final RunCommand owner;
   String get _templateFileName => owner.argResults![owner.kOptionTemplate];
 
+  /// Reads the template file with the matching name of the -t option.
   TemplateDef readTemplate() {
     final filePath = '${owner.config.templateDirectory}$_templateFileName.yaml';
     final templateContent = read(filePath).toParagraph();