+++ lib/common/yaml_utils.dart
@@ -2,6 +2,7 @@ import 'package:autobot/common/exceptions.dart';
 import 'package:yaml/yaml.dart';
 
 extension YamlUtil on YamlMap {
+  /// Returns the value of the given [key] or throws [MissingYamlField].
   dynamic require(String key, {required String fileName}) {
     final value = this[key];