+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.5.0
+- Support for nested mustache variabels
+- Allwos for adding shell scripts to the script section
+- Adds the `input-file` option argument for reading variables from a yaml file
 ## 0.4.0
 - Allow javascript section in template
 ## 0.3.0+2